+++ src/DefaultConsumer.php
@@ -13,7 +13,6 @@
 namespace PMG\Queue;
 
 use Psr\Log\LoggerInterface;
-use Psr\Log\NullLogger;
 
 /**
  * A default implementation of `Consumer`. Runs jobs via an executor.
@@ -21,9 +20,8 @@ use Psr\Log\NullLogger;
  * @since   2.0
  * @api
  */
-final class DefaultConsumer implements Consumer
+final class DefaultConsumer extends AbstractConsumer
 {
-    const EXIT_ERROR = 2;
 
     /**
      * @var Driver
@@ -40,60 +38,16 @@ final class DefaultConsumer implements Consumer
      */
     private $retries;
 
-    /**
-     * @var LoggerInterface
-     */
-    private $logger;
-
-    /**
-     * @var boolean
-     */
-    private $running = false;
-
-    /**
-     * @var int
-     */
-    private $exitCode = 0;
-
     public function __construct(
         Driver $driver,
         MessageExecutor $executor,
         RetrySpec $retries=null,
         LoggerInterface $logger=null
     ) {
+        parent::__construct($logger);
         $this->driver = $driver;
         $this->executor = $executor;
         $this->retries = $retries ?: new Retry\LimitedSpec();
-        $this->logger = $logger ?: new NullLogger();
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function run($queueName)
-    {
-        $this->running = true;
-        while ($this->running) {
-            try {
-                $this->once($queueName);
-            } catch (Exception\MustStop $e) {
-                $this->logger->warning('Caught a must stop exception, exiting: {msg}', [
-                    'msg'   => $e->getMessage(),
-                ]);
-                $this->stop();
-                $this->exitCode = $e->getCode();
-            } catch (\Exception $e)  {
-                // likely means means something went wrong with the driver
-                $this->logger->emergency('Caught an unexpected {cls} exception, exiting: {msg}', [
-                    'cls' => get_class($e),
-                    'msg' => $e->getMessage(),
-                ]);
-                $this->stop();
-                $this->exitCode = self::EXIT_ERROR;
-            }
-        }
-
-        return $this->exitCode;
     }
 
     /**
@@ -109,27 +63,19 @@ final class DefaultConsumer implements Consumer
         $result = false;
         $message = $envelope->unwrap();
 
-        $this->logger->debug('Executing message {msg}', ['msg' => $message->getName()]);
+        $this->getLogger()->debug('Executing message {msg}', ['msg' => $message->getName()]);
         $result = $this->executeMessage($message);
-        $this->logger->debug('Executed message {msg}', ['msg' => $message->getName()]);
+        $this->getLogger()->debug('Executed message {msg}', ['msg' => $message->getName()]);
 
         if ($result) {
             $this->driver->ack($queueName, $envelope);
-            $this->logger->debug('Acknowledged message {msg}', ['msg' => $message->getName()]);
+            $this->getLogger()->debug('Acknowledged message {msg}', ['msg' => $message->getName()]);
         } else {
             $this->failed($queueName, $envelope);
-            $this->logger->debug('Failed message {msg}', ['msg' => $message->getName()]);
+            $this->getLogger()->debug('Failed message {msg}', ['msg' => $message->getName()]);
         }
 
         return $result;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function stop()
-    {
-        $this->running = false;
     }
 
     private function failed($queueName, Envelope $env)
@@ -154,7 +100,7 @@ final class DefaultConsumer implements Consumer
             // below. We only log here because we can't make guarantees about
             // the implementation of the executor and whether or not it actually
             // throws exceptions on failure (see ForkingExecutor).
-            $this->logger->critical('Unexpected {cls} exception handling {name} message: {msg}', [
+            $this->getLogger()->critical('Unexpected {cls} exception handling {name} message: {msg}', [
                 'cls'   => get_class($e),
                 'name'  => $message->getName(),
                 'msg'   => $e->getMessage()